feat: set liglicko2 old state on create game
